Moved get setting from shop/helpers to settings/helpers along with imports
