+++ shopyo/modules/box__default/settings/helpers.py
@@ -1,0 +1,19 @@
+from modules.box__default.settings.models import Settings
+
+
+def get_setting(name):
+    """
+    Used as key-value lookup from Settings table
+
+    Parameters
+    ----------
+    name: str
+        name of key
+
+    Returns
+    -------
+    str
+        value of key
+    """
+    s = Settings.query.get(name)
+    return s.value
